Cleanup in prep for working on styling changes
This is a cleanup of some unused classes and testing how to style the maker descriptions in preperation for the beging changes to they style of the page making it responsive and user friendly for all that use it. Cleaned up the unused class "item" in the different divs. This only occured in the North Carolina files as this will be the test for the rest of the maps